DDCNLS-332: Moved DES case classes into data/des package and into separate files. Added tests for EDH connector fractions happy path and fractions controller happy path.
